+++ pph_in_C/headers/sparce_vector.h
@@ -8,7 +8,7 @@
  */
 
 typedef struct _vector_index {
-    unsigned long int     pos; /* index which the vector is not zero */
+    unsigned long long int     pos; /* index which the vector is not zero */
     struct _vector_index *next;
 }vector_index;
 
@@ -18,16 +18,16 @@ typedef struct {
 
 vector  *alloc_vec               (void);
 
-void    add_index_to_vector      (vector*, unsigned long int);
+void    add_index_to_vector      (vector*, unsigned long long int);
 
-void    remove_index_from_vector (vector*, unsigned long int);
+void    remove_index_from_vector (vector*, unsigned long long int);
 
 boolean is_this_vector_zero      (vector*);
 
 void    sum_these_vectors        (vector*, vector*);
 
 /* vector with 1 on position k and 0 otherwise */
-vector  *I_k                     (unsigned long int); 
+vector  *I_k                     (unsigned long long int); 
 
 void    print_vec_nicely         (vector*, char*);
