+++ pph_in_C/src/definitions.c
@@ -4,7 +4,7 @@
 
 boolean are_these_regular_paths_the_same (regular_path path1, regular_path path2, dim_path path_dim) {
 
-    unsigned int i;
+    unsigned long long int i;
     for (i = 0; i <= path_dim; ++i)
         if ( path1[i] != path2[i] ) return FALSE;
 
@@ -12,7 +12,7 @@ boolean are_these_regular_paths_the_same (regular_path path1, regular_path path2
 }
 
 boolean is_this_path_a_regular_path (regular_path path, dim_path path_dim) {
-    unsigned int i;
+    unsigned long long int i;
 
     for (i = 0; i < path_dim; ++i)
         if( path[i] == path[i+1] ) return FALSE;
